requirements.txt
black
discord.py
